extracting the scrollToBottom logic to a hook
